requirements.txt
torch==2.1.1
torchaudio==2.1.1
pypinyin==0.42.0
nltk==3.6.3
g2p_en==2.1.0
g2p_id_py==0.3.1
pyopenjtalk==0.3.3
g2pK==0.9.3
ToJyutping==0.2.1
uroman-python==1.2.8.1
pydub==0.25.1
matplotlib==3.7.3
PyYAML
typeguard==2.13.3
textgrid
transformers
pandas
scipy
accelerate==0.24.1
demucs==4.0.1
loguru==0.7.2